requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
websockets
python-multipart
jinja2
sqladmin
python-dotenv
gunicorn
pymysql
cryptography
